Finishing version

+++ src/main/java/org/harunjaganjac/example/controllers/EmployeeController.java
@@ -1,6 +1,7 @@
 package org.harunjaganjac.example.controllers;
 
 import org.harunjaganjac.example.models.Employee;
+import org.harunjaganjac.example.response.EmployeeResponse;
 import org.harunjaganjac.example.services.EmployeeProjectService;
 import org.harunjaganjac.example.services.EmployeeService;
 
@@ -14,20 +15,27 @@ public final class EmployeeController {
         this.employeeService = employeeService;
         this.employeeProjectService = employeeProjectService;
     }
-    public boolean createEmployee(Employee newEmployee) {
+    public EmployeeResponse createEmployee(Employee newEmployee) {
         return employeeService.createEmployee(newEmployee);
     }
     public Employee getEmployee(String id) {
         return employeeService.getEmployeeById(id);
     }
-    public boolean updateEmployee(String id, Employee updatedEmployee) {
+    public EmployeeResponse updateEmployee(String id, Employee updatedEmployee) {
         return employeeService.updateEmployee(id, updatedEmployee);
     }
     public boolean deleteEmployee(String id) {
-        return employeeService.deleteEmployee(id);
+        var result= employeeProjectService.removeEmployee(id);
+        if(result){
+            return employeeService.deleteEmployee(id);
+        }
+        return false;
     }
     public List<Employee> getAllEmployees() {
         return employeeService.getAllEmployees();
+    }
+    public List<String> getAllEmployeeIds() {
+        return employeeService.getAllEmployeeIds();
     }
     public boolean assignProject(String employeeId, String projectId) {
         return employeeProjectService.assignProject(employeeId, projectId);
